+++ Realm/RLMObject.mm
@@ -22,10 +22,7 @@
 #import "RLMObjectStore.hpp"
 #import "RLMQueryUtil.hpp"
 #import "RLMUtil.hpp"
-
-#if REALM_SWIFT
-#import <Realm/Realm-Swift.h>
-#endif
+#import "RLMSwiftSupport.h"
 
 #import <objc/runtime.h>
 
@@ -191,11 +188,9 @@
 // overridden at runtime per-class for performance
 + (NSString *)className {
     NSString *className = NSStringFromClass(self);
-#if REALM_SWIFT
     if ([RLMSwiftSupport isSwiftClassName:className]) {
         className = [RLMSwiftSupport demangleClassName:className];
     }
-#endif
     return className;
 }